+++ Icepack/IcepackIgnoreAttribute.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace Icepack
+{
+    /// <summary> Marks a field to be ignored during serialization/deserialization. </summary>
+    public class IcepackIgnoreAttribute : Attribute { }
+}