+++ Icepack/IgnorePropertyAttribute.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace Icepack
-{
-    /// <summary> Marks a field to be ignored during serialization/deserialization via Icepack. </summary>
-    public class IgnorePropertyAttribute : Attribute { }
-}