+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.0.2
+
+* Support for serializing structs, hashsets, and dictionaries.
+* Allow marking reference fields to be serialized as value-only.
+* Fixed a bug when serializing multiple references to the same object.
+* All tests passing.
+
 # v0.0.1
 
 * Basic functionality in place, including serialization/deserialization, type registering, and support for arrays and lists.